+++ frontend/src/pages/MyStreams.js
@@ -4,11 +4,86 @@ import useSWR from 'swr';
 import {fromNear, fromTaras} from '../components/Helpers';
 import {StreamControls} from '../features/stream-control/StreamControls';
 import {NEAR, loader, TARAS} from '../components/Helpers';
+import {CategoryTabs, Filter, FilterOptionWithCounter} from '../components/kit';
+import {StreamCard} from '../components/StreamCard';
+
+const STREAM_STATUS = {
+  ARCHIVED: 'ARCHIVED',
+  ACTIVE: 'ACTIVE',
+  PAUSED: 'PAUSED',
+};
+const STREAM_DIRECTION = {
+  IN: 'in',
+  OUT: 'out',
+};
+const STREAM_TYPE_FILTER = {
+  ALL: 'All',
+  INCOMING: 'Incoming',
+  OUTGOING: 'Outgoing',
+};
+
+const STREAM_STATUS_FILTER = {
+  ALL: 'All',
+  ACTIVE: 'Active',
+  ARCHIVED: 'Archived',
+};
+
+function useFilters({items, filters}) {
+  console.log('UseFilters', {items, filters});
+  return items.filter((item) => filters.every((filter) => filter(item)));
+}
+
+function useStreamFilters(streams) {
+  const [statusFilterValue, setStatusFilter] = useState(
+    STREAM_STATUS_FILTER.ACTIVE,
+  );
+  const [directionFilterValue, setDirectionFilter] = useState(
+    STREAM_TYPE_FILTER.ALL,
+  );
+
+  const statusFilters = {
+    [STREAM_STATUS_FILTER.ALL]: () => true,
+    [STREAM_STATUS_FILTER.ACTIVE]: (stream) =>
+      stream.status === STREAM_STATUS.ACTIVE,
+    [STREAM_STATUS_FILTER.ARCHIVED]: (stream) =>
+      stream.status === STREAM_STATUS.ARCHIVED,
+  };
+
+  const directionFilters = {
+    [STREAM_TYPE_FILTER.ALL]: () => true,
+    [STREAM_TYPE_FILTER.INCOMING]: (stream) =>
+      STREAM_DIRECTION.IN === stream.direction,
+    [STREAM_TYPE_FILTER.OUTGOING]: (stream) =>
+      STREAM_DIRECTION.OUT === stream.direction,
+  };
+
+  const filteredItems = useFilters({
+    items: streams,
+    filters: [
+      directionFilters[directionFilterValue],
+      statusFilters[statusFilterValue],
+    ],
+  });
+
+  return {
+    directionFilterValue,
+    statusFilterValue,
+    setDirectionFilter,
+    setStatusFilter,
+    filteredItems,
+  };
+}
 
 export function MyStreamsPage() {
   const near = useNear();
   const [token, setToken] = useState('NEAR');
-  const [showButtons, setShowButtons] = useState(true);
+
+  const isIncomingStream = (stream) => {
+    if (stream.owner_id === near.near.accountId) {
+      return false;
+    }
+    return true;
+  };
 
   const {data: account} = useSWR(
     ['account', near.near.accountId],
@@ -18,19 +93,15 @@ export function MyStreamsPage() {
     },
   );
 
-  const outputs = account && account.outputs ? account.outputs : [];
-  console.log('accout', account);
-  console.log('outputs', outputs);
+  const inputs = (account && account.inputs) || [];
+  inputs.forEach((input) => (input.direction = 'in'));
+  const outputs = (account && account.outputs) || [];
+  outputs.forEach((output) => (output.direction = 'out'));
 
-  const depositList = account
-    ? outputs.map((output, id) => {
-        return (
-          <option key={id} value={id}>
-            {output.stream_id}
-          </option>
-        );
-      })
-    : [];
+  const allStreams = inputs.concat(outputs);
+
+  const filter = useStreamFilters(allStreams);
+  console.log('ACCOUNT', account);
 
   function selectClick(e) {
     var selectBox = document.getElementById('selectBox');
@@ -40,7 +111,6 @@ export function MyStreamsPage() {
   }
   async function depositClick(e) {
     e.preventDefault();
-    setShowButtons(false);
     var selectBox = document.getElementById('selectBox');
     if (selectBox.selectedIndex > 0) {
       const stream = outputs[selectBox.selectedIndex - 1];
@@ -79,167 +149,115 @@ export function MyStreamsPage() {
         console.log('deposit TARAS res', res);
       }
     }
-    setShowButtons(true);
   }
 
-  const outputsTable = outputs.map((output, id) => {
-    const history = output.history.map((record, id) => {
-      return (
-        <div className="d-flex flex-row justify-content-between w-100">
-          <div>{record.actor}</div>
-          <div>{record.action_type}</div>
-          {record.amount ? (
-            <div>{fromNear(record.amount).toFixed(2)}</div>
-          ) : (
-            <div />
-          )}
-          <div>{record.timestamp}</div>
-        </div>
-      );
-    });
-    return (
-      <div
-        className="card"
-        style={{width: '90%', margin: '15px', backgroundColor: '#141414'}}
-        key={id}
-      >
-        <div className="card-body">
-          <div className="d-flex flex-row justify-content-between w-100">
-            <div className="col-2 m-1">{output.receiver_id}</div>
-            <small className="col-1 m-1">
-              <small>
-                <samp className="text-secondary">
-                  {output.stream_id.substr(0, 6)}...
-                </samp>
-              </small>
-            </small>
-            <small className="col-1 m-1">
-              {output.token_name === 'NEAR'
-                ? fromNear(output.balance).toFixed(2)
-                : fromTaras(output.balance).toFixed(2)}
-            </small>
-            <small className="col-1 m-1">{output.token_name}</small>
-            <small className="col-1 m-1">
-              {output.token_name === 'NEAR'
-                ? (fromNear(output.tokens_per_tick) * 1e9).toFixed(2)
-                : (fromTaras(output.tokens_per_tick) * 1e9).toFixed(2)}{' '}
-              {output.token_name}/s
-            </small>
-            <small className="col-1 m-1">
-              {output.token_name === 'NEAR'
-                ? fromNear(output.available_to_withdraw).toFixed(2)
-                : fromTaras(output.available_to_withdraw).toFixed(2)}
-            </small>
-            <small className="col-1 m-1">
-              {output.token_name === 'NEAR'
-                ? fromNear(output.tokens_transferred).toFixed(2)
-                : fromTaras(output.tokens_transferred).toFixed(2)}
-            </small>
-            <small className="col-1 m-1">{output.status}</small>
-            <div className="col-2 m-1">
-              <StreamControls output={output} />
-            </div>
-          </div>
-          <div className="h2">History</div>
-          <strong className="d-flex flex-row justify-content-between w-100">
-            <div>Actor</div>
-            <div>Action</div>
-            <div>Amount</div>
-            <div>Timestamp</div>
-          </strong>
-          {history}
-        </div>
-      </div>
-    );
-  });
+  // const outputsTable = outputs.map((output, id) => {
+  //   return (
+  //     <div
+  //       className="card"
+  //       style={{width: '90%', margin: '15px', backgroundColor: '#141414'}}
+  //       key={id}
+  //     >
+  //       <div className="card-body">
+  //         <div className="d-flex flex-row justify-content-between w-100">
+  //           <div className="col-2 m-1">{output.receiver_id}</div>
+  //           <small className="col-1 m-1">
+  //             <small>
+  //               <samp className="text-secondary">
+  //                 {output.stream_id.substr(0, 6)}...
+  //               </samp>
+  //             </small>
+  //           </small>
+  //           <small className="col-1 m-1">
+  //             {output.token_name === 'NEAR'
+  //               ? fromNear(output.balance).toFixed(2)
+  //               : fromTaras(output.balance).toFixed(2)}
+  //           </small>
+  //           <small className="col-1 m-1">{output.token_name}</small>
+  //           <small className="col-1 m-1">
+  //             {output.token_name === 'NEAR'
+  //               ? (fromNear(output.tokens_per_tick) * 1e9).toFixed(2)
+  //               : (fromTaras(output.tokens_per_tick) * 1e9).toFixed(2)}{' '}
+  //             {output.token_name}/s
+  //           </small>
+  //           <small className="col-1 m-1">
+  //             {output.token_name === 'NEAR'
+  //               ? fromNear(output.tokens_available).toFixed(2)
+  //               : fromTaras(output.tokens_available).toFixed(2)}
+  //           </small>
+  //           <small className="col-1 m-1">
+  //             {output.token_name === 'NEAR'
+  //               ? fromNear(output.tokens_transferred).toFixed(2)
+  //               : fromTaras(output.tokens_transferred).toFixed(2)}
+  //           </small>
+  //           <small className="col-1 m-1">{output.status}</small>
+  //           <div className="col-2 m-1">
+  //             <StreamControls output={output} />
+  //           </div>
+  //         </div>
+  //       </div>
+  //     </div>
+  //   );
+  // });
 
   return (
-    <div className="twind-container">
-      <div>
-        <h4>Streams owned by you</h4>
-        <div
-          className="card"
-          style={{width: '90%', margin: '15px', backgroundColor: '#181818'}}
-        >
-          <div className="card-body">
-            <div className="d-flex flex-row justify-content-between w-100">
-              <div className="col-2 m-1">Receiver</div>
-              <div className="col-1 m-1">Stream ID</div>
-              <div className="col-1 m-1">Tokens left</div>
-              <div className="col-1 m-1">Token name</div>
-              <div className="col-1 m-1">Stream speed</div>
-              <div className="col-1 m-1">Tokens unlocked</div>
-              <div className="col-1 m-1">Tokens transferred</div>
-              <div className="col-1 m-1">Stream status</div>
-              <div className="col-2 m-1" />
-            </div>
-          </div>
-        </div>
-        {outputsTable}
-      </div>
+    <div className="twind-container twind-p-12">
+      <h1 className="twind-text-3xl twind-mb-12">All Streams</h1>
+      <div className="twind-flex">
+        <Filter
+          className="twind-mr-5"
+          options={Object.values(STREAM_TYPE_FILTER)}
+          label="Type:"
+          active={filter.directionFilterValue}
+          onChange={filter.setDirectionFilter}
+          renderOption={(option) => {
+            const countMap = {
+              [STREAM_TYPE_FILTER.INCOMING]: inputs.length,
+              [STREAM_TYPE_FILTER.OUTGOING]: outputs.length,
+              [STREAM_TYPE_FILTER.ALL]: inputs.length + outputs.length,
+            };
 
-      <div
-        className="card"
-        style={{width: '90%', margin: '15px', backgroundColor: '#141414'}}
-      >
-        <div className="card-body">
-          <h5 className="card-title">Deposit tokens</h5>
-          <h6 className="card-subtitle mb-3 text-muted">
-            Extend your existing stream
-          </h6>
-          <form onSubmit={(e) => depositClick(e)}>
-            <div className="form-group mb-2">
-              <label htmlFor="ownerInput" className="mb-2">
-                Stream
-              </label>
-              <select
-                id="selectBox"
-                className="form-control custom-select"
-                onChange={(e) => selectClick(e)}
-              >
-                <option selected>Open this select menu</option>
-                {depositList}
-              </select>
-            </div>
-            <label htmlFor="depositTokensInput" className="mb-2">
-              Amount of tokens to deposit
-            </label>
-            <div className="input-group mb-2">
-              <div className="input-group-prepend">
-                <span
-                  className="input-group-text"
-                  id="basic-addon2"
-                  style={{
-                    backgroundColor: '#303030',
-                    color: '#e0e0e0',
-                    fontSize: '150%',
-                  }}
-                >
-                  {token === 'NEAR' ? NEAR : TARAS}
-                </span>
-              </div>
-              <input
-                className="form-control"
-                id="depositTokensInput"
-                placeholder="0.03"
-                describedby="basic-addon2"
-                style={{backgroundColor: '#101010', color: '#e0e0e0'}}
+            return (
+              <FilterOptionWithCounter
+                option={option}
+                count={countMap[option]}
               />
-            </div>
+            );
+          }}
+        />
+        <Filter
+          options={Object.values(STREAM_STATUS_FILTER)}
+          label="Status:"
+          active={filter.statusFilterValue}
+          onChange={filter.setStatusFilter}
+          renderOption={(option) => {
+            const countMap = {
+              [STREAM_STATUS_FILTER.ACTIVE]: allStreams.filter(
+                (stream) => stream.status === STREAM_STATUS.ACTIVE,
+              ).length,
+              [STREAM_STATUS_FILTER.ALL]: allStreams.length,
+              [STREAM_STATUS_FILTER.ARCHIVED]: allStreams.filter(
+                (stream) => stream.status === STREAM_STATUS.ARCHIVED,
+              ).length,
+            };
 
-            {near.inited && showButtons ? (
-              <button
-                disabled={!near.auth.signedIn}
-                className="btn btn-primary mt-4"
-              >
-                Deposit
-              </button>
-            ) : (
-              <div className="mt-4">{loader()}</div>
-            )}
-          </form>
-        </div>
+            return (
+              <FilterOptionWithCounter
+                option={option}
+                count={countMap[option]}
+              />
+            );
+          }}
+        />
       </div>
+
+      {filter.filteredItems.map((stream) => (
+        <StreamCard
+          stream={stream}
+          className="twind-mb-4"
+          direction={isIncomingStream(stream) ? 'in' : 'out'}
+        />
+      ))}
     </div>
   );
 }